Adapt for new neovim-deps branch.
The neovim-deps Launchpad branch is imported from neovim/deps on GitHub.
Dependency sources are in a subdirectory and can't be moved into the correct
place by the recipe alone (`run` instructions are not allowed in recipes on
Launchpad).

Also add a workaround to explicitly generate the `./configure` script for
jemalloc, which is not committed to neovim/deps as it's in jemalloc's
`.gitignore` file.